+++ src/psyclone/psyir/transformations/intrinsics/minval2loop_trans.py
@@ -33,138 +33,138 @@
 # -----------------------------------------------------------------------------
 # Author: R. W. Ford, STFC Daresbury Lab
 
-'''Module providing a transformation from a PSyIR MAXVAL intrinsic to
-an equivalent PSyIR loop structure. This could be useful if the MAXVAL
+'''Module providing a transformation from a PSyIR MINVAL intrinsic to
+an equivalent PSyIR loop structure. This could be useful if the MINVAL
 operator is not supported by the back-end, the required
 parallelisation approach, or if the performance in the inline code is
 better than the intrinsic.
 
 '''
 from psyclone.psyir.nodes import Reference, IntrinsicCall
-from psyclone.psyir.transformations.intrinsics.mms_base_trans import (
-    MMSBaseTrans)
+from psyclone.psyir.transformations.intrinsics.array_reduction_base_trans \
+    import ArrayReductionBaseTrans
 
 
-class Maxval2CodeTrans(MMSBaseTrans):
-    '''Provides a transformation from a PSyIR MAXVAL IntrinsicCall node to
+class Minval2LoopTrans(ArrayReductionBaseTrans):
+    '''Provides a transformation from a PSyIR MINVAL IntrinsicCall node to
     an equivalent PSyIR loop structure that is suitable for running in
     parallel on CPUs and GPUs. Validity checks are also performed.
 
-    If MAXVAL contains a single positional argument which is an array,
-    the maximum value of all of the elements in the array is returned
+    If MINVAL contains a single positional argument which is an array,
+    the minimum value of all of the elements in the array is returned
     in the the scalar R.
 
     .. code-block:: fortran
 
-        R = MAXVAL(ARRAY)
+        R = MINVAL(ARRAY)
 
     For example, if the array is two dimensional, the equivalent code
     for real data is:
 
     .. code-block:: fortran
 
-        R = TINY(R)
+        R = HUGE(R)
         DO J=LBOUND(ARRAY,2),UBOUND(ARRAY,2)
           DO I=LBOUND(ARRAY,1),UBOUND(ARRAY,1)
-            R = MAX(R, ARRAY(I,J))
+            R = MIN(R, ARRAY(I,J))
 
     If the mask argument is provided then the mask is used to
-    determine whether the maxval is applied:
+    determine whether the minval is applied:
 
     .. code-block:: fortran
 
-        R = MAXVAL(ARRAY, mask=MOD(ARRAY, 2.0)==1)
+        R = MINVAL(ARRAY, mask=MOD(ARRAY, 2.0)==1)
 
     If the array is two dimensional, the equivalent code
     for real data is:
 
     .. code-block:: fortran
 
-        R = TINY(R)
+        R = HUGE(R)
         DO J=LBOUND(ARRAY,2),UBOUND(ARRAY,2)
           DO I=LBOUND(ARRAY,1),UBOUND(ARRAY,1)
             IF (MOD(ARRAY(I,J), 2.0)==1) THEN
-              R = MAX(R, ARRAY(I,J))
+              R = MIN(R, ARRAY(I,J))
 
     The dimension argument is currently not supported and will result
     in a TransformationError exception being raised.
 
     .. code-block:: fortran
 
-        R = MAXVAL(ARRAY, dimension=2)
+        R = MINVAL(ARRAY, dimension=2)
 
-    The array passed to MAXVAL may use any combination of array
+    The array passed to MINVAL may use any combination of array
     syntax, array notation, array sections and scalar bounds:
 
     .. code-block:: fortran
 
-        R = MAXVAL(ARRAY) ! array syntax
-        R = MAXVAL(ARRAY(:,:)) ! array notation
-        R = MAXVAL(ARRAY(1:10,lo:hi)) ! array sections
-        R = MAXVAL(ARRAY(1:10,:)) ! mix of array section and array notation
-        R = MAXVAL(ARRAY(1:10,2)) ! mix of array section and scalar bound
+        R = MINVAL(ARRAY) ! array syntax
+        R = MINVAL(ARRAY(:,:)) ! array notation
+        R = MINVAL(ARRAY(1:10,lo:hi)) ! array sections
+        R = MINVAL(ARRAY(1:10,:)) ! mix of array section and array notation
+        R = MINVAL(ARRAY(1:10,2)) ! mix of array section and scalar bound
 
-    An example use of this transformation is given below:
+    For example:
 
     >>> from psyclone.psyir.backend.fortran import FortranWriter
     >>> from psyclone.psyir.frontend.fortran import FortranReader
-    >>> from psyclone.psyir.transformations import Maxval2CodeTrans
-    >>> code = ("subroutine maxval_test(array)\\n"
+    >>> from psyclone.psyir.transformations import Minval2LoopTrans
+    >>> code = ("subroutine minval_test(array)\\n"
     ...         "  real :: array(10,10)\\n"
     ...         "  real :: result\\n"
-    ...         "  result = maxval(array)\\n"
+    ...         "  result = minval(array)\\n"
     ...         "end subroutine\\n")
     >>> psyir = FortranReader().psyir_from_source(code)
     >>> sum_node = psyir.children[0].children[0].children[1]
-    >>> Maxval2CodeTrans().apply(sum_node)
+    >>> Minval2LoopTrans().apply(sum_node)
     >>> print(FortranWriter()(psyir))
-    subroutine maxval_test(array)
+    subroutine minval_test(array)
       real, dimension(10,10) :: array
       real :: result
       integer :: idx
       integer :: idx_1
     <BLANKLINE>
-      result = TINY(result)
+      result = HUGE(result)
       do idx = 1, 10, 1
         do idx_1 = 1, 10, 1
-          result = MAX(result, array(idx_1,idx))
+          result = MIN(result, array(idx_1,idx))
         enddo
       enddo
     <BLANKLINE>
-    end subroutine maxval_test
+    end subroutine minval_test
     <BLANKLINE>
 
     '''
-    _INTRINSIC_NAME = "MAXVAL"
-    _INTRINSIC_TYPE = IntrinsicCall.Intrinsic.MAXVAL
+    _INTRINSIC_NAME = "MINVAL"
+    _INTRINSIC_TYPE = IntrinsicCall.Intrinsic.MINVAL
 
     def _loop_body(self, lhs, rhs):
-        '''Provide the body of the nested loop that computes the maximum value
+        '''Provide the body of the nested loop that computes the minimum value
         of the lhs and rhs.
 
-        :param lhs: the lhs value for the max operation.
+        :param lhs: the lhs value for the min operation.
         :type lhs: :py:class:`psyclone.psyir.nodes.Node`
-        :param rhs: the rhs value for the max operation.
+        :param rhs: the rhs value for the min operation.
         :type rhs: :py:class:`psyclone.psyir.nodes.Node`
 
-        :returns: a MAX IntrinsicCall.
+        :returns: a MIN IntrinsicCall.
         :rtype: :py:class:`psyclone.psyir.nodes.IntrinsicCall`
 
         '''
-        # return max(lhs,rhs)
-        return IntrinsicCall.create(IntrinsicCall.Intrinsic.MAX, [lhs, rhs])
+        # return min(lhs,rhs)
+        return IntrinsicCall.create(IntrinsicCall.Intrinsic.MIN, [lhs, rhs])
 
     def _init_var(self, var_symbol):
-        '''The initial value for the variable that computes the maximum value
+        '''The initial value for the variable that computes the minimum value
         of an array.
 
         :param var_symbol: the symbol used to store the final result.
         :type var_symbol: :py:class:`psyclone.psyir.symbols.DataSymbol`
 
         :returns: PSyIR for the value to initialise the variable that
-            computes the maximum value.
+            computes the minimum value.
         :rtype: :py:class:`psyclone.psyir.nodes.IntrinsicCall`
 
         '''
         return IntrinsicCall.create(
-            IntrinsicCall.Intrinsic.TINY, [Reference(var_symbol)])
+            IntrinsicCall.Intrinsic.HUGE, [Reference(var_symbol)])
